Add city selection ui (activity+presenter), fix weather dto types, cache - singleton

+++ app/src/main/java/com/ikholopov/yamblz/weather/weathermobilization/data/dto/weather/WeatherInfo.java
@@ -2,7 +2,7 @@ package com.ikholopov.yamblz.weather.weathermobilization.data.dto.weather;
 
 public class WeatherInfo {
     public float temp;
-    public int pressure;
+    public float pressure;
     public int humidity;
     public float temp_min;
     public float temp_max;
